+++ packages/arb-token-bridge-ui/src/components/TransferPanel/TransferPanelUtils.ts
@@ -1,6 +1,7 @@
 import * as Sentry from '@sentry/react'
 
 import { isUserRejectedError } from '../../util/isUserRejectedError'
+import { useArbQueryParams } from '../../hooks/useArbQueryParams'
 
 export enum ImportTokenModalStatus {
   // "IDLE" is here to distinguish between the modal never being opened, and being closed after a user interaction
@@ -23,5 +24,27 @@ export function getWarningTokenDescription(warningTokenType: number) {
 export function onTxError(error: unknown) {
   if (!isUserRejectedError(error)) {
     Sentry.captureException(error)
+  }
+}
+
+export function useTokenFromSearchParams(): {
+  tokenFromSearchParams: string | undefined
+  setTokenQueryParam: (token: string | undefined) => void
+} {
+  const [{ token: tokenFromSearchParams }, setQueryParams] = useArbQueryParams()
+
+  const setTokenQueryParam = (token: string | undefined) =>
+    setQueryParams({ token })
+
+  if (!tokenFromSearchParams) {
+    return {
+      tokenFromSearchParams: undefined,
+      setTokenQueryParam
+    }
+  }
+
+  return {
+    tokenFromSearchParams,
+    setTokenQueryParam
   }
 }
